requirements.txt
streamlit
google-cloud-speech
fpdf2
requests
numpy
sounddevice
google-auth